pyproject.toml
[tool.poetry]
name = "panel"
version = "0.1.0"
description = ""
authors = ["Professional <pyostr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.14"
django = "^5.2.7"
python-dotenv = "^1.1.1"
pytelegrambotapi = "^4.29.1"
pymysql = "^1.1.2"
boto3 = "^1.40.50"
django-storages = "^1.14.6"
redis = "^6.4.0"
celery = "^5.5.3"
django-ninja = "^1.4.3"
django-ninja-extra = "^0.30.1"
cryptography = "^46.0.2"
whitenoise = "^6.11.0"
gunicorn = "^23.0.0"
pillow = "^11.3.0"
channels-redis = "^4.3.0"
pika = "^1.3.2"
django-celery-beat = "^2.8.1"
django-cors-headers = "^4.9.0"
django-json-widget = "^2.0.4"
pyotp = "^2.9.0"
qrcode = "^8.2"
twisted = "^25.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
